requirements.txt
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.25.0
Pillow>=10.0.0
XlsxWriter>=3.1.0